Create new MongoClient Bean only created when we are in MONGO mode to fix initialisation issues.
